Only send a SIGHUP if the pid is greater than -1, otherwise all PIDs greater than -1 will get the SIGHUP... and that is bad. (closes issue #11453) Reported by: alanmcmillan
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@92199 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ res/res_agi.c
@@ -2208,7 +2208,8 @@ static enum agi_result run_agi(struct ast_channel *chan, char *request, AGI *agi
 		if (needhup) {
 			needhup = 0;
 			dead = 1;
-			kill(pid, SIGHUP);
+			if (pid > -1)
+				kill(pid, SIGHUP);
 		}
 		ms = -1;
 		c = ast_waitfor_nandfds(&chan, dead ? 0 : 1, &agi->ctrl, 1, NULL, &outfd, &ms);
